replaced BigDecimal with Double

+++ src/main/java/io/contek/invoker/binancelinear/api/common/_Account.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 @NotThreadSafe
@@ -13,17 +12,17 @@ public class _Account {
   public boolean canWithdraw;
   public long updateTime;
   public boolean multiAssetsMargin;
-  public BigDecimal totalInitialMargin;
-  public BigDecimal totalMaintMargin;
-  public BigDecimal totalWalletBalance;
-  public BigDecimal totalUnrealizedProfit;
-  public BigDecimal totalMarginBalance;
-  public BigDecimal totalPositionInitialMargin;
-  public BigDecimal totalOpenOrderInitialMargin;
-  public BigDecimal totalCrossWalletBalance;
-  public BigDecimal totalCrossUnPnl;
-  public BigDecimal availableBalance;
-  public BigDecimal maxWithdrawAmount;
+  public Double totalInitialMargin;
+  public Double totalMaintMargin;
+  public Double totalWalletBalance;
+  public Double totalUnrealizedProfit;
+  public Double totalMarginBalance;
+  public Double totalPositionInitialMargin;
+  public Double totalOpenOrderInitialMargin;
+  public Double totalCrossWalletBalance;
+  public Double totalCrossUnPnl;
+  public Double availableBalance;
+  public Double maxWithdrawAmount;
   public List<_AccountAsset> assets;
   public List<_AccountPosition> positions;
 }
