+++ src/main/java/io/contek/invoker/binancelinear/api/common/_AccountPosition.java
@@ -1,24 +1,23 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _AccountPosition {
 
   public String symbol;
-  public BigDecimal initialMargin;
-  public BigDecimal maintMargin;
-  public BigDecimal unrealizedProfit;
-  public BigDecimal positionInitialMargin;
-  public BigDecimal openOrderInitialMargin;
-  public BigDecimal leverage;
+  public Double initialMargin;
+  public Double maintMargin;
+  public Double unrealizedProfit;
+  public Double positionInitialMargin;
+  public Double openOrderInitialMargin;
+  public Double leverage;
   public Boolean isolated;
-  public BigDecimal entryPrice;
-  public BigDecimal maxNotional;
-  public BigDecimal bidNotional;
-  public BigDecimal askNotional;
+  public Double entryPrice;
+  public Double maxNotional;
+  public Double bidNotional;
+  public Double askNotional;
   public String positionSide;
-  public BigDecimal positionAmt;
+  public Double positionAmt;
   public long updateTime;
 }
