+++ src/main/java/io/contek/invoker/binancelinear/api/common/_FeeRates.java
@@ -1,12 +1,11 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _FeeRates {
 
   public String symbol;
-  public BigDecimal makerCommissionRate;
-  public BigDecimal takerCommissionRate;
+  public Double makerCommissionRate;
+  public Double takerCommissionRate;
 }
