+++ src/main/java/io/contek/invoker/binancelinear/api/common/_MiniTickerSummary.java
@@ -1,22 +1,21 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _MiniTickerSummary {
 
     public String symbol;
-    public BigDecimal priceChange;
-    public BigDecimal priceChangePercent;
-    public BigDecimal weightedAvgPrice;
-    public BigDecimal lastPrice;
-    public BigDecimal lastQty;
-    public BigDecimal openPrice;
-    public BigDecimal highPrice;
-    public BigDecimal lowPrice;
-    public BigDecimal volume;
-    public BigDecimal quoteVolume;
+    public Double priceChange;
+    public Double priceChangePercent;
+    public Double weightedAvgPrice;
+    public Double lastPrice;
+    public Double lastQty;
+    public Double openPrice;
+    public Double highPrice;
+    public Double lowPrice;
+    public Double volume;
+    public Double quoteVolume;
     public long openTime;
     public long closeTime;
     public long firstId;
