+++ src/main/java/io/contek/invoker/binancelinear/api/common/_Order.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.binancelinear.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Order {
@@ -9,17 +8,17 @@ public class _Order {
   public String symbol;
   public String orderId;
   public String clientOrderId;
-  public BigDecimal price;
-  public BigDecimal avgPrice;
-  public BigDecimal origQty;
-  public BigDecimal executedQty;
+  public Double price;
+  public Double avgPrice;
+  public Double origQty;
+  public Double executedQty;
   public String status;
   public String timeInForce;
   public String type;
   public boolean reduceOnly;
   public String side;
   public String positionSide;
-  public BigDecimal stopPrice;
+  public Double stopPrice;
   public long updateTime;
   public String workingType;
   public long traceNano = System.nanoTime();
