+++ src/main/java/io/contek/invoker/binancelinear/api/websocket/user/AccountUpdateChannel.java
@@ -6,7 +6,6 @@ import io.contek.invoker.binancelinear.api.websocket.user.constants.UserEventTyp
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 @ThreadSafe
@@ -50,21 +49,21 @@ public final class AccountUpdateChannel extends UserWebSocketChannel<AccountUpda
   public static final class BalanceUpdate {
 
     public String a; // asset
-    public BigDecimal wb; // wallet balance
-    public BigDecimal cw; // cross wallet balance
-    public BigDecimal bc; // Balance Change except PnL and Commission
+    public Double wb; // wallet balance
+    public Double cw; // cross wallet balance
+    public Double bc; // Balance Change except PnL and Commission
   }
 
   @NotThreadSafe
   public static final class PositionUpdate {
 
     public String s; // symbol
-    public BigDecimal pa; // position amount
-    public BigDecimal ep; // entry price
-    public BigDecimal cr; // pre-fee accumulated realized
-    public BigDecimal up; // unrealized PnL
+    public Double pa; // position amount
+    public Double ep; // entry price
+    public Double cr; // pre-fee accumulated realized
+    public Double up; // unrealized PnL
     public String mt; // margin type
-    public BigDecimal iw; // isolated wallet
+    public Double iw; // isolated wallet
     public String ps; // position side
   }
 }
